chore: move controller utils interface and fix  issue
