+++ source/pages/Home/Home.tsx
@@ -8,14 +8,8 @@ import { TxsPanel } from './TxsPanel';
 
 export const Home = () => {
   const controller = getController();
-  const {
-    networks,
-    activeNetwork,
-    fiat,
-    activeAccount,
-    activeToken,
-    lastLogin,
-  } = useStore();
+  const { networks, activeNetwork, fiat, activeAccount, lastLogin } =
+    useStore();
   const [symbol, setSymbol] = useState('SYS');
   const [fiatPriceValue, setFiatPriceValue] = useState('');
 
@@ -24,59 +18,39 @@ export const Home = () => {
   const { navigate } = useUtils();
 
   const isSyscoinChain = Boolean(networks.syscoin[activeNetwork.chainId]);
+  const chain = isSyscoinChain ? 'syscoin' : 'ethereum';
 
   const balance = isSyscoinChain
     ? activeAccount.balances.syscoin
     : activeAccount.balances.ethereum;
 
   const setChainSymbol = async () => {
-    const chain = isSyscoinChain ? 'syscoin' : 'ethereum';
     const symbol = await getSymbolByChain(chain);
+
+    setSymbol(symbol);
 
     return symbol;
   };
 
   const getFiatPrice = async () => {
-    const value = await getFiatAmount(
+    const amount = await getFiatAmount(
       balance || 0,
-      isSyscoinChain ? '' : activeToken,
       4,
-      String(fiat.asset)
+      String(fiat.asset).toUpperCase()
     );
 
-    return value;
+    setFiatPriceValue(String(amount));
+
+    return amount;
   };
 
   const isUnlocked =
     controller.wallet.isUnlocked() && activeAccount.address !== '';
 
   useEffect(() => {
-    let isMounted = true;
-
-    setChainSymbol().then((response: any) => {
-      if (isMounted) {
-        setSymbol(response);
-      }
-    });
-
-    return () => {
-      isMounted = false;
-    };
-  }, [isUnlocked]);
-
-  useEffect(() => {
-    let isMounted = true;
-
-    getFiatPrice().then((response: any) => {
-      if (isMounted) {
-        setFiatPriceValue(response);
-      }
-    });
-
-    return () => {
-      isMounted = false;
-    };
-  }, [activeNetwork]);
+    setChainSymbol();
+    getFiatPrice();
+  }, [isUnlocked, activeNetwork]);
 
   const isSysTestnet = activeNetwork.chainId === 5700;
   const symbolByChain = isSysTestnet ? 'tsys' : symbol;
