+++ source/scripts/Background/controllers/index.ts
@@ -1,4 +1,6 @@
-import ControllerUtils, { IControllerUtils } from './ControllerUtils';
+import { IControllerUtils } from 'types/controllers';
+
+import ControllerUtils from './ControllerUtils';
 import MainController from './MainController';
 
 export interface IMasterController {
